fix: rename ModalField to Modal
