+++ frontend/src/components/general/fileUpload/FileUpload.tsx
@@ -3,7 +3,8 @@ import axios from "axios";
 import { Check, MoveLeft, MoveRight, Trash2 } from "lucide-react";
 import { useAuthHeader } from "@/hooks/useAuthHeader";
 import PrimaryButton from "../buttons/PrimaryButton";
-import ModalField from "../form/ModalField";
+import ModalField from "../form/Modal";
+
 
 
 interface FileUploadProps {
